add methods and tests for getProposal and getProposals
