+++ components/proposals/components/ProposalStepSelect.tsx
@@ -3,7 +3,7 @@ import { useMemo } from 'react';
 import { TagSelect } from 'components/common/BoardEditor/components/properties/TagSelect/TagSelect';
 import type { PropertyValueDisplayType } from 'components/common/BoardEditor/interfaces';
 import type { IPropertyOption } from 'lib/focalboard/board';
-import type { ProposalWithUsersLite } from 'lib/proposal/interface';
+import type { ProposalWithUsersLite } from 'lib/proposal/getProposals';
 
 import { useBatchUpdateProposalStatusOrStep } from '../hooks/useBatchUpdateProposalStatusOrStep';
