+++ components/proposals/hooks/useProposals.tsx
@@ -7,7 +7,7 @@ import { useGetProposalsBySpace } from 'charmClient/hooks/proposals';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { usePages } from 'hooks/usePages';
 import { useWebSocketClient } from 'hooks/useWebSocketClient';
-import type { ProposalWithUsersLite } from 'lib/proposal/interface';
+import type { ProposalWithUsersLite } from 'lib/proposal/getProposals';
 import type { UpdateProposalRequest } from 'lib/proposal/updateProposal';
 import type { WebSocketPayload } from 'lib/websockets/interfaces';
