+++ lib/proposal/mapDbProposalToProposal.ts
@@ -9,18 +9,10 @@ import type {
 } from '@charmverse/core/prisma';
 import type { ProposalEvaluation } from '@charmverse/core/prisma-client';
 import { getCurrentEvaluation } from '@charmverse/core/proposals';
-import { sortBy } from 'lodash';
 
 import { getProposalFormFields } from 'lib/proposal/form/getProposalFormFields';
 
-import { getCurrentStep } from './getCurrentStep';
-import type {
-  ProposalFields,
-  PopulatedEvaluation,
-  ProposalWithUsersAndRubric,
-  ProposalWithUsersLite,
-  TypedFormField
-} from './interface';
+import type { ProposalFields, PopulatedEvaluation, ProposalWithUsersAndRubric, TypedFormField } from './interface';
 
 type FormFieldsIncludeType = {
   form: {
@@ -80,53 +72,6 @@ export function mapDbProposalToProposal({
           id: form.id
         }
       : null
-  };
-
-  return proposalWithUsers;
-}
-
-// used for mapping data for proposal blocks/tables which dont need all the evaluation data
-export function mapDbProposalToProposalLite({
-  proposal,
-  permissions
-}: {
-  proposal: Proposal & {
-    authors: ProposalAuthor[];
-    evaluations: (ProposalEvaluation & { reviewers: ProposalReviewer[] })[];
-    rewards: { id: string }[];
-  };
-  permissions?: ProposalPermissionFlags;
-}): ProposalWithUsersLite {
-  const { rewards, ...rest } = proposal;
-  const currentEvaluation = getCurrentEvaluation(proposal.evaluations);
-  const fields = (rest.fields as ProposalFields) ?? null;
-
-  const proposalWithUsers: ProposalWithUsersLite = {
-    id: rest.id,
-    createdBy: rest.createdBy,
-    authors: proposal.authors,
-    archived: proposal.archived || undefined,
-    formId: rest.formId || undefined,
-    spaceId: rest.spaceId,
-    evaluations: sortBy(proposal.evaluations, 'index').map((e) => ({
-      title: e.title,
-      index: e.index,
-      type: e.type,
-      result: e.result,
-      id: e.id
-    })),
-    permissions,
-    currentStep: getCurrentStep({
-      evaluations: proposal.evaluations,
-      hasPendingRewards: (fields?.pendingRewards ?? []).length > 0,
-      proposalStatus: proposal.status,
-      hasPublishedRewards: rewards.length > 0
-    }),
-    currentEvaluationId: proposal.status !== 'draft' && proposal.evaluations.length ? currentEvaluation?.id : undefined,
-    status: proposal.status,
-    reviewers: (proposal.status !== 'draft' && currentEvaluation?.reviewers) || [],
-    rewardIds: rewards.map((r) => r.id),
-    fields
   };
 
   return proposalWithUsers;
