+++ pages/api/proposals/[id]/index.ts
@@ -1,13 +1,12 @@
 import { prisma } from '@charmverse/core/prisma-client';
-import { getCurrentEvaluation } from '@charmverse/core/proposals';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
 
 import { ActionNotPermittedError, NotFoundError, onError, onNoMatch, requireKeys, requireUser } from 'lib/middleware';
 import { permissionsApiClient } from 'lib/permissions/api/client';
 import { canAccessPrivateFields } from 'lib/proposal/form/canAccessPrivateFields';
+import { getProposal } from 'lib/proposal/getProposal';
 import type { ProposalWithUsersAndRubric } from 'lib/proposal/interface';
-import { mapDbProposalToProposal } from 'lib/proposal/mapDbProposalToProposal';
 import type { UpdateProposalRequest } from 'lib/proposal/updateProposal';
 import { updateProposal } from 'lib/proposal/updateProposal';
 import { withSessionRoute } from 'lib/session/withSession';
@@ -20,77 +19,23 @@ handler.get(getProposalController).use(requireUser).put(updateProposalController
 
 async function getProposalController(req: NextApiRequest, res: NextApiResponse<ProposalWithUsersAndRubric>) {
   const proposalId = req.query.id as string;
-  const userId = req.session.user?.id;
+  const userId = req.session.user?.id as string | undefined;
 
-  const proposal = await prisma.proposal.findUnique({
-    where: {
-      id: proposalId
-    },
-    include: {
-      evaluations: {
-        orderBy: {
-          index: 'asc'
-        },
-        include: {
-          permissions: true,
-          reviewers: true,
-          rubricCriteria: {
-            orderBy: {
-              index: 'asc'
-            }
-          },
-          rubricAnswers: true,
-          draftRubricAnswers: true,
-          vote: true
-        }
-      },
-      authors: true,
-      page: { select: { id: true, sourceTemplateId: true, type: true } },
-      reviewers: true,
-      rewards: true,
-      form: {
-        include: {
-          formFields: {
-            orderBy: {
-              index: 'asc'
-            }
-          }
-        }
-      }
-    }
-  });
-
-  if (!proposal) {
-    throw new NotFoundError();
-  }
   const permissionsByStep = await permissionsApiClient.proposals.computeAllProposalEvaluationPermissions({
     // Proposal id is the same as page
-    resourceId: proposal?.id,
+    resourceId: proposalId,
     userId
   });
+  // If we are viewing a proposal template, we can see all private fields since the user might be creating a proposal
+  const canAccessPrivateFormFields = await canAccessPrivateFields({ userId, proposalId });
 
-  const currentEvaluation = getCurrentEvaluation(proposal.evaluations);
+  const proposal = await getProposal({ id: proposalId, canAccessPrivateFormFields, permissionsByStep });
 
-  const currentPermissions =
-    proposal.status === 'draft'
-      ? permissionsByStep.draft
-      : currentEvaluation && permissionsByStep[currentEvaluation.id];
-
-  if (!currentPermissions?.view) {
+  if (!proposal.permissions?.view) {
     throw new NotFoundError();
   }
 
-  // If we are viewing a proposal template, we can see all private fields since the user might be creating a proposal
-  const canAccessPrivateFormFields = await canAccessPrivateFields({ proposal, userId, proposalId: proposal.id });
-
-  return res.status(200).json(
-    mapDbProposalToProposal({
-      proposal,
-      permissions: currentPermissions,
-      permissionsByStep,
-      canAccessPrivateFormFields
-    })
-  );
+  return res.status(200).json(proposal);
 }
 
 async function updateProposalController(req: NextApiRequest, res: NextApiResponse) {
