--- pages/api/proposals/[id]/is-reviewer.ts
@@ -1,28 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from 'next';
-import nc from 'next-connect';
-
-import { ActionNotPermittedError, onError, onNoMatch, requireUser } from 'lib/middleware';
-import { permissionsApiClient } from 'lib/permissions/api/client';
-import { withSessionRoute } from 'lib/session/withSession';
-
-const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
-
-handler.use(requireUser).get(getReviewerIds);
-
-async function getReviewerIds(req: NextApiRequest, res: NextApiResponse<boolean>) {
-  const proposalId = req.query.id as string;
-  const userId = req.session.user.id;
-
-  const permissions = await permissionsApiClient.proposals.computeProposalPermissions({
-    resourceId: proposalId,
-    userId
-  });
-
-  if (!permissions.view) {
-    throw new ActionNotPermittedError(`You cannot view this proposal`);
-  }
-
-  return res.status(200).send(permissions.review || permissions.evaluate);
-}
-
-export default withSessionRoute(handler);
